pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "anima-dlrmamba"
version = "0.1.0"
description = "ANIMA DLRMamba module: low-rank SS2D + structure-aware distillation for RGB-IR detection"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "AIFLOW LABS" }
]
dependencies = [
  "torch>=2.2.0",
  "pillow>=10.0.0",
  "numpy>=1.26.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/anima_dlrmamba"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
